requirements.txt
requests
pytest
voluptuous
pytest-voluptuous
selene==2.0.0b16
allure-pytest
python-dotenv
